jupiter-1 Research spike: schema validation of xAPI payloads
From some perfunctory tests, I think the z-schema library is pretty solid. I am going to use it to create a new schema starting with fugu13's schema to generate something that complies with draft 4 of the json-schema specification and successfully validates the 1.0.0 version of the xAPI specification.

If I am successful, I will port z-schema to Java. Ideally, the schema will be able to be verified using Jackson's streaming API, but that may be a tall order for this initial revision.